Move FakeAnsibleRepo to Spec::Support
